Overall security improvements and code optimizations

+++ TestMe/Controllers/TestQuestionsController.cs
@@ -44,11 +44,12 @@ namespace TestMe.Controllers
 
         public async Task<IActionResult> Index(int? id)
         {
-            if (id == null)
+            if (id is null)
                 return NotFound();
 
             var test = await _testingPlatform.TestManager
                 .FindAsync(t => t.AppUserId == _userId && t.Id == id);
+
             if (test is null)
                 return NotFound();
 
@@ -57,12 +58,13 @@ namespace TestMe.Controllers
 
         public async Task<IActionResult> Details(int? id)
         {
-            if (id == null)
+            if (id is null)
                 return NotFound();
 
             var testQuestion = await _testingPlatform.TestQuestionManager
                 .FindAsync(tq => tq.AppUserId == _userId && tq.Id == id);
-            if (testQuestion == null)
+
+            if (testQuestion is null)
                 return NotFound();
 
             return View(testQuestion);
@@ -70,12 +72,12 @@ namespace TestMe.Controllers
 
         public async Task<IActionResult> Create(int? id)
         {
-            if (id == null)
+            if (id is null)
                 return NotFound();
 
             var test = await _testingPlatform.TestManager
                 .FindAsync(t => t.AppUserId == _userId && t.Id == id);
-            if (test == null)
+            if (test is null)
                 return NotFound();
 
             if (!(test.TestCode is null))
@@ -89,6 +91,12 @@ namespace TestMe.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Create([Bind("QuestionText, TestId, PreformattedText")] TestQuestion testQuestion)
         {
+            var test = await _testingPlatform.TestManager
+                .FindAsync(t => t.AppUserId == _userId && t.Id == testQuestion.TestId);
+
+            if (test is null)
+                return NotFound();
+
             if (ModelState.IsValid)
             {
                 testQuestion.AppUserId = _userId;
@@ -96,11 +104,12 @@ namespace TestMe.Controllers
                 return RedirectToAction(nameof(Index), new { id = testQuestion.TestId });
             }
 
-            var test = await _testingPlatform.TestManager
+            test = await _testingPlatform.TestManager
                 .FindAsync(t => t.AppUserId == _userId && t.Id == testQuestion.TestId);
-            if (test == null)
+
+            if (test is null)
                 return NotFound();
-            
+
             if (!(test.TestCode is null))
                 return NotFound();
 
@@ -109,12 +118,12 @@ namespace TestMe.Controllers
 
         public async Task<IActionResult> Edit(int? id)
         {
-            if (id == null)
+            if (id is null)
                 return NotFound();
 
             var testQuestion = await _testingPlatform.TestQuestionManager
                 .FindAsync(t => t.AppUserId == _userId && t.Id == id);
-            if (testQuestion == null)
+            if (testQuestion is null)
                 return NotFound();
 
             if (!(testQuestion.Test.TestCode is null))
@@ -128,6 +137,11 @@ namespace TestMe.Controllers
         public async Task<IActionResult> Edit(int id, [Bind("Id, QuestionText, TestId, PreformattedText")] TestQuestion testQuestion)
         {
             if (id != testQuestion.Id)
+                return NotFound();
+
+            var test = await _testingPlatform.TestManager
+                .FindAsync(t => t.AppUserId == _userId && t.Id == testQuestion.TestId);
+            if (test is null)
                 return NotFound();
 
             if (ModelState.IsValid)
@@ -156,11 +170,11 @@ namespace TestMe.Controllers
 
         public async Task<IActionResult> Delete(int? id)
         {
-            if (id == null)
+            if (id is null)
                 return NotFound();
 
             var testQuestion = await _testingPlatform.TestQuestionManager.FindAsync(t => t.AppUserId == _userId && t.Id == id);
-            if (testQuestion == null)
+            if (testQuestion is null)
                 return NotFound();
             
             if (!(testQuestion.Test.TestCode is null))
@@ -174,6 +188,9 @@ namespace TestMe.Controllers
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
             var testQuestion = await _testingPlatform.TestQuestionManager.FindAsync(t => t.AppUserId == _userId && t.Id == id);
+            if (testQuestion is null)
+                return NotFound();
+
             var testId = testQuestion.TestId;
             foreach (var testAnswer in testQuestion.TestAnswers.Where(ta => !(ta.ImageName is null)))
                 _testingPlatform.AnswerImageManager.DeleteAnswerImage(testAnswer.ImageName);
